dubai-property-ai-2/api/requirements.txt
google-generativeai>=0.5.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
resend
